add query support to urlAssembly logic
